Support loading and querying const tables
As the first part of the pandas-like API, this patch adds support
for loading Python Iterables to Greenplum as "constant tables"
and querying them. This is implemented by using the VALUES
clause in SQL. This is the simplest Table that I can think of.

This patch also adds some minimal test cases using pytest.