Support function calls on constants
Being able to call functions is essential for data analytics. This
patch adds support for calling functions in Greenplum. Given
that Column is not completed when the work starts, this patch
only implements function calls on constants but not on
tables. Changes are required to support function calls on tables.