Support creating UDFs from Python functions
It is common that a data scientist needs to define new UDFs for
complex analytics tasks. This patch introduces a new decorator that
converts a Python function into a Greenplum UDF. And the created
UDF can be called in exactly the same way as ordinary Python
functions. This makes it much easier to work with Greenplum UDFs.