Support Table Join: (#49)
Join is essential when manipulating several tables for data analysts
or data scientists. This patch supports different types of JOIN in
Greenplum: inner, left, right, full, cross, and natural. Each type of join
has each function for better use of IDE.
The join is done on columns by giving the join condition, it can be 
equal or not. We can also give a list of columns to select columns
that we want to join.