Support creating and calling array functions
Array functions takes arrays as arguments. Before calling the
function, each attribute in the rows will be aggregated as an
array and the array will be passed to the function as an
argument. In this way, the function can have access to the
attributes in all rows. This makes it possible for the user to do
complex analytics and for the PL handler to optimize the
performance