Support passing Python objects and type casts (#60)
It is very common that an Expr contains Python objects. This patch
adds a wrapper to convert a Python object to a SQL coonstant
systematically. To ensrue that the type of the const matches
what's desired, this patch also adds support for type casts by
"calling" instances of the Type class.

After adding the Type class, columns now can have types. This
opens the door to column type inference.