Allow passing column names directly to `join()`
Preiously, `Table.join()` accepts generic `Expr`s as arguments. This
often requires the user to refer to the table again and again, which
can be annoying and error-prone.

This patch simplies join() by changing its arguments to column
names. The user can do renaming if necessary by passing dicts of
names. This makes join() less powerful but easier to use.