Fix error when calling same UDF again (#99)
Previously, a UDF is not created until it is called. This would cause
error when calling it more than once since the same function has
already existed.

This patch fixes the error by introducing a global dict to keep track
of which functions are defined and functions with the same name
cannot be defined more than once.

Moreover, this patch removes support for creating non-temp
functions and types. This helps make the Python script easier to
reproduce by data scientists.