Add functional API
GreenplymPython's goal is to help user break a big SQL query into
several easy function calls in Python. However, this allows the user
to contruct incorrect queries. For example, the user can write an
`Expr` composed of columns from `Table`s from more than one
database. But such `Expr` cannot be evaluated by the server.

This patch tries to mitigate this issue by adding a functional API.
This makes it possible to delay binding `Expr`s and other objects
to `Table`s and `Database`s until the we know how to do it
automatically for the user. This also makes the API easier to
compose, and the entire program concise and fun to write.