Refactor how `Expr`s are related to `Table` (#107)
Previously, there are multiple ways Exprs are related to a Table, 
such as by assigning an Expr to a new column, or by `Expr.to_table()`
explicitly, etc. This makes the interface complicated and hard to
learn.

This patch solves the issue by unifying all these ways to using
`assign()` and removing all the others. With `assign()`, the user
can bind any Expr to a new column with name specified explicitly.
This makes the API simple and clear, and also making it convenient
for us to implement table column inference later. 

Co-authored-by: Ruxue Zeng <zeng_ruxue@126.com>