Support DISTINCT
This patch adds support for DISTINCT, both SELECT DISTINCT and
distinct-qualified aggregation like count(DISTINCT). Specifically,
distinct() are added to both `Table` and `AggregateFunction` as
methods, respectively.

This patch also fixes error when assign() is empty after group_by(),
changing it to truly return a `Table`, as specified in the signature.