Update docstring for next beta release (#109)
Complete and clear documentation is necessary for the user to understand
how a product works. This patch updates the documentation of GreenplumPython
with the latest feature modifications.

* Resupport `apply()` function (#110)

When applying a function on the table, the user can only select returning columns,
the assign() function keeps all of the columns of the table applied, which could bring
extra work to users. This patch supports apply() function which returns a Table
only with returning columns of UDF.

Co-authored-by: Xuebin Su <sxuebin@vmware.com>