Support `in` operator (#120)
Membership testing is essential in data analytics. Previously, due to
lacking of the `in` operator, users have to use `join()` followed by
`distinct()` to get the correct result, which can be inefficient.

This patch adds the `in` operator, enabling using semi-join for
efficient membership testing, but with some limitations:
- Nested `in` exprs, e.g. `t["a"].in_(t2.["b"].in_(t3["c"]))` is not
supported. And only one table other than the current is allowed.
- `in` exprs can only be used to form operator exprs but no others
such as function calls.