Feat: set up to display sql or not (#119)
To understand the result of GreenplumPython, users could
want to investigate the SQL query sent by GreenplumPython
to the database. This patch adds a global method to set up the option
"sql_on".
If "sql_on" is True, GreenplumPython will print the query in the output,
otherwise not. The default value is False.