Fix error when table in non-default schema
When retrieving data of a table in a non-default schema with `to_json()`,
an error will occur. This is because the SQL parser treats schema name
as table name and table name as column name and finds the table with
the schema name does not exist.

This patch fixes the error by renaming the table with the `AS` clause
to avoid ambiguity.