Fix lifetime issue for functions (#124)
Previously, to avoid double creation error, we maintianed a global scope
of functions and throwed an error if the name of a new function
conflicts with those in the global scope. However, this makes it very
inconvenient if the user want to change the function definition, because
they will need to restart the process, i.e. the IPython kernel if they
are using Jupyter Notebook in that case.

This patch removes the global scope so that the user can make the change
by re-creating the function object. To avoid double creation error, we
force all UDFs to be TEMP objects so that

> An object is alive in database system if and only if both the Python
> object and the db connection are alive.

This is the rule of reasoning the livetime of an object, including
tables, functions, operators, and types, unless the object is saved
explicitly by the user.

Co-authored-by: Ruxue Zeng <zeng_ruxue@126.com>