Fix error when table in non-default schema (#131)
When retrieving data of a table in a non-default schema with
`to_json()`,
an error will occur. This is because the SQL parser treats schema name
as table name and table name as column name and finds the table with
the schema name does not exist.

This patch fixes the error by renaming the table with the `AS` clause
to avoid ambiguity.

This patch also adds a test case for converting pandas DataFame to
GreenplumPython Table without introducing pandas as dependency.