Refactor how DataFrame is formatted
This patch fixes type errors in the formatting method of a DataFrame.
Those errors are previously ignored.

This patch also adds row numbers to the formatted string so that
DataFrame can be formatted even when it is empty.

Moreover, this patch also makes Row of DataFrame looks more like
immutable dict.