Support outside dependencies for functions
Previously, to create a UDF in database, the user is required to write
all the dependencies inside the function body. This often results in
annoying boilerplate code

This patch adds support for refering to objects outside of the function,
including functions and modules. This closes the gap between a UDF
and a normal Python function.