Change psycopg2 package to binary (#147)
As a dependency of GreenplumPython, psycopg2 is a source package. To
install it, the client side
need to have
    * libpq and its C header files,
    * C compilers, such as gcc

installed so that the C code in the psycopg2 package can be compiled.
This can be complicated
to handle different system environment, to sidestep this problem, we
change our dependency to
psycopg2-binary, which is a binary package.

This patch adds also more examples and explaination in documentation.
And change the lowest python
version requirement to Python3.9.