Support outside dependencies for UDFs (#142)
Previously, to create a UDF in database, the user is required to write
all the dependencies inside the function body. This often results in
annoying boilerplate code

This patch adds support for refering to objects outside of the UDF,
including functions and modules. This closes the gap between a UDF
and a normal Python function.

This is achieved by using a pickling library, dill, to serialize a
Python function as well as all of its dependencies before sending it 
to server. However, it requires to install the library on server to
deserialize the function before executing it. If the library is not 
usable, we will fallback automatically to the "source code" solution
by also sending the source code of the function to the server.

---------

Co-authored-by: yihong0618 <zouzou0208@gmail.com>