Simplify API for database connection (#167)
Previously, function for connecting to a db has a custom interface.
This can cause confusion to newcommers. 

This patch simplifies the interface of the function so that it
confronts to widely adopted conventions. Now users are able to learn
the usage by referring to the doc of libpq or psycopg2.

---------

Co-authored-by: Ruxue Zeng <36695415+Quasimewdo@users.noreply.github.com>