Make db a parameter of serialize() (#208)
Previously, `db` is an attribute of class `Expr` and its value
cannot be decided when init a new object. As a result, when the
`Expr` is serialized, we need to explicitly bind it to a db to
create functions or types. However, in some cases we forgot to
do that, which lead to errros.

To avoid forgotting this, this patch moves db to a parameter of
`serialize()` and pass db to it each time we call it.

This patch fixes the following bugs in `serialize()`:

1. Database is not found when trying to create a function or data
   type.
2. Star notation `*` is not serialized correctly in column fileds
   when expanding a composite type result.
3. `count(*)` returns wrong result when dataframe contains `None`.

---------

Co-authored-by: Ruxue Zeng <zeng_ruxue@126.com>