Add module for indexing and searching embeddings (#207)
Vector embeddings are essential for searching unstructured data, such
as images and texts. Typically, to implement semantic search with
embeddings, user will need to go through a complex process:

1. Generate embeddings with transformers;
2. Save the generated embeddings into a table;
3. Create index on the generated embeddings;
4. Search for items with most similar embeddings.

To simplify the process, this patch adds a new module that provides
only two functions: `create_index()` and `search()`. This hides all
the embedding-related details and enables user to focus only on the
text or image data that they are really interested in.

---------

Co-authored-by: Ruxue Zeng <zeng_ruxue@126.com>