Fix missing dependency on segments
Previsously, to record dependency, we used `UPDATE` in SQL to update
pg_depend. However, this only took effect on coordinator but not on
segments. As a result, the embedding index cannot be drop on segments
even with `DROP CASCADE`.

This patch fixes this issue by calling the C function
`recordDependencyOn()` with ctypes on all segements including
coordinator. As a result, `DROP CASCADE` will also clean the embedding
index on segments.