Support bulk loading from client
Previously, to upload data from client to the database, the only way is
to call `DataFrame.from_rows()` or `DataFrame.from_columns()`. In
this way, data will be serialized as part of the query. When data size
is large, the query will become very long and can take large amount
of memory.

This patch fixes the issue by uploading the data files to the database
with `COPY FROM STDIN`. In this way, the query can be kept short.
After the data files are copied, a UDF is called to parse the files into
rows. This makes it much easier to support various amount of data
formats, inluding the unseen ones.