Support installing packages from client
On servers that have no Internet access, it is hard to install the
required Python packages for a UDF.

To mitigate this issue, this patch enables installing packages from
the client side. This works as follows:

1. Download the required packages to a directory on client with
  `pip download`
2. Archive and upload the downloaded packages to server
3. Install the uploaded packages with `pip install --find-links`.

Please note that this only install the packages on a single server
host. To make the packages available to other hosts, an NFS
might be helpful.