Add tutorial for installing packages on server
This patch adds a notebook on how to install Python packages on server
without Internet access. In the tutorial, NFS is used to share the
packages amoung multiple hosts in the same cluster. This can simplify
the installation and maintenance process in a distributed env.

This patch also adds warnings to notify that some features are currently
experimental.