Fix error when `/tmp` is a symlink
On macOS, `/tmp` is a symlink to `/private/tmp`. When creating the
tarball, the symlink will be resolved, making the file paths on server
are different from those on client. This results in a
`FileNotFoundError`.

This patch fixes the issue by resolving the local path to the real one
if it is a symlink.