Install pgvector on CI (#225)
Previsously we did not have test for embedding indexing. This patch
installs [pgvector](https://github.com/pgvector/pgvector) from the 
PostgreSQL's apt repo so that we can test the embedding-related
features on CI.

To keep it simple, instead of modifying a container at runtime, this
patch adds a new `tox` env, called `test-container`, which builds the
docker image of database server with all dependencies before running
the tests. The `test` env is still available for running tests with
an existing server.

To make tests more effective, this patch also adds an option called 
`server_use_pickler` to `pytest`, indicating that whether to use the
pickler `dill` to deserialize UDFs on server. As a result, we can
have test cases for the pickler, which will fail when the pickler is
not applicable. Such test cases are currently missing and we have
encountered several issues because of that.