Fix missing unique key when searching embeddings (#226)
Unique key is required when searching embeddings to join the
embedding table and the original data table before returning the
results. Previously, search on a dataframe that created from an
existing table in database failed due to lacking of unique key in the
dataframe.

This patch fixes the issue by recording the unique key when
`create_index()` in `pg_class` so that the info can be read when
`search()`.