Add doc for server requirements (#227)
As more (experimental) new features are added, it is getting more and
more complex to setup a server to meet the requirements.

To make the process easier, this patch adds a page `requirements.rst`,
to specify the requirements for client and for server. It is for all
the features including the experimental ones.

This patch also makes minor adjustments to make the interface of the
experimental features more clear and easier to work with.