Fix unique key missing problem with search (#232)
When creating embeddings, we store the unique key attnum in pg_options. 
Previously, the unique key attnum was taken from the original table, so
the error occurs if the
embedding table and the original table have a different number or order
of columns.
This patch saves the unique key attnum from the embedding table instead.